modified reset password design
