+++ src/Components/Settings/SettingView.tsx
@@ -1,47 +1,14 @@
-import { Form, Formik } from 'formik'
 import React, { useState } from 'react'
-import { FaRegEye, FaRegEyeSlash } from 'react-icons/fa'
-import { useNavigate } from 'react-router'
-import { OnChangeCallback } from 'react-toastify'
-import * as Yup from 'yup'
 import AccessManagement from '../../Components/Settings/AccessManagement'
 import Toast from '../../Components/Toast'
-import Button from '../Button'
-import InputField from '../Input/Input'
+import ResetPassword from './ResetPassword'
 
 const SettingView = () => {
   const [activeTab, setActiveTab] = useState('accessManagement')
-  const [loading] = useState(false)
-  const [showPassword, setShowPassword] = useState(false)
-  const [showNewPassword, setShowNewPassword] = useState(false)
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false)
   const [pin, setPin] = useState<string[]>(new Array(6).fill(''))
   const [pin2, setPin2] = useState<string[]>(new Array(6).fill(''))
   const [isFirstPinSet, setIsFirstPinSet] = useState(false)
   const [toast, setToast] = useState(false)
-
-  const navigate = useNavigate()
-
-  const initialValues = {
-    password: '',
-    newPassword: '',
-    confrimPassword: '',
-  }
-
-  const validationSchema = Yup.object().shape({
-    password: Yup.string()
-      .required('Password cannot be empty')
-      .max(20, 'Password must not exceed 20 characters')
-      .trim(),
-    newPassword: Yup.string()
-      .required('New Password cannot be empty')
-      .max(20, 'New Password must not exceed 20 characters')
-      .trim(),
-    confrimPassword: Yup.string()
-      .required('Confirm Password cannot be empty')
-      .max(20, 'Confirm Password must not exceed 20 characters')
-      .trim(),
-  })
 
   const handleBackspace = (e: React.KeyboardEvent, index: number) => {
     if (e.key === 'Backspace' && !pin[index] && index > 0) {
@@ -87,12 +54,6 @@ const SettingView = () => {
     setTimeout(() => {}, 1000)
   }
 
-  const handleChangeSubmit = () => {
-    setTimeout(() => {
-      navigate('/signin')
-    }, 1000)
-  }
-
   const handleChangeAuthCall = () => {
     setTimeout(() => {
       //   navigate("/signin");
@@ -127,7 +88,7 @@ const SettingView = () => {
             Access Management
           </button>
           {/* not integrated yet */}
-          {/* <button
+          <button
             className={`px-4 py-2 text-md  ${
               activeTab === 'password' ? 'font-bold' : 'font-normal'
             }  ${
@@ -150,29 +111,14 @@ const SettingView = () => {
             onClick={() => setActiveTab('authPin')}
           >
             Change authentication pin
-          </button> */}
+          </button>
           {/* not integrated yet */}
         </div>
 
         {/* Tab Content */}
         <div className="mt-10">
           {activeTab === 'accessManagement' && <AccessManagement />}
-          {activeTab === 'password' && (
-            <PasswordReset
-              validationSchema={validationSchema}
-              showPassword={showPassword}
-              showNewPassword={showNewPassword}
-              showConfirmPassword={showConfirmPassword}
-              setShowPassword={() => setShowPassword(!showPassword)}
-              setShowNewPassword={() => setShowNewPassword(!showNewPassword)}
-              setShowConfirmPassword={() =>
-                setShowConfirmPassword(!showConfirmPassword)
-              }
-              initialValues={initialValues}
-              loading={loading}
-              handleSubmitChange={handleChangeSubmit}
-            />
-          )}
+          {activeTab === 'password' && <ResetPassword />}
           {activeTab === 'authPin' && (
             <ChangeAuthPin
               pin={pin}
@@ -193,102 +139,6 @@ const SettingView = () => {
 }
 
 export default SettingView
-
-// Password Reset Form
-interface PasswordResetProps {
-  initialValues: any
-  handleSubmitChange: OnChangeCallback
-  validationSchema: any
-  showPassword: boolean
-  setShowPassword: Function
-  showNewPassword: boolean
-  setShowNewPassword: Function
-  showConfirmPassword: boolean
-  setShowConfirmPassword: Function
-  loading: boolean
-}
-
-function PasswordReset({
-  initialValues,
-  handleSubmitChange,
-  validationSchema,
-  showPassword,
-  setShowPassword,
-  showNewPassword,
-  setShowNewPassword,
-  showConfirmPassword,
-  setShowConfirmPassword,
-  loading,
-}: PasswordResetProps) {
-  return (
-    <div className="w-full  min-h-2 flex justify-center">
-      <div className="px-6 py-10 mx-auto w-2/4 bg-white shadow-lg rounded-xl">
-        <div className="">
-          <h1 className="text-2xl font-bold mb-2">Password reset</h1>
-          <p className="text-gray-600">Change your password</p>
-        </div>
-        <Formik
-          initialValues={initialValues}
-          validateOnChange={true}
-          validateOnBlur={true}
-          onSubmit={handleSubmitChange}
-          validationSchema={validationSchema}
-        >
-          {({ isValid, dirty, setFieldValue, setFieldTouched }) => (
-            <Form>
-              <div className="mt-8">
-                <InputField
-                  label=""
-                  name="Existing password"
-                  type={showPassword ? 'password' : 'text'}
-                  placeholder="Existing password"
-                  onClick={() => setShowPassword(!showPassword)}
-                  icon={showPassword ? <FaRegEye /> : <FaRegEyeSlash />}
-                  setFieldValue={setFieldValue}
-                  setFieldTouched={setFieldTouched}
-                />
-              </div>
-              <div className="mt-8">
-                <InputField
-                  label=""
-                  name="New password"
-                  type={showNewPassword ? 'password' : 'text'}
-                  placeholder="New password"
-                  onClick={() => setShowNewPassword(!showNewPassword)}
-                  icon={showNewPassword ? <FaRegEye /> : <FaRegEyeSlash />}
-                  setFieldValue={setFieldValue}
-                  setFieldTouched={setFieldTouched}
-                />
-              </div>
-
-              <div className="mt-8">
-                <InputField
-                  label=""
-                  name="Confirm new password"
-                  type={showConfirmPassword ? 'password' : 'text'}
-                  placeholder="Confirm new password"
-                  onClick={() => setShowConfirmPassword(!showConfirmPassword)}
-                  icon={showConfirmPassword ? <FaRegEye /> : <FaRegEyeSlash />}
-                  setFieldValue={setFieldValue}
-                  setFieldTouched={setFieldTouched}
-                />
-              </div>
-              <div className="mt-3">
-                <Button
-                  text="Save changes"
-                  loading={loading}
-                  active={true}
-                  bg_color="#007A61"
-                  text_color="#FFFFFF"
-                />
-              </div>
-            </Form>
-          )}
-        </Formik>
-      </div>
-    </div>
-  )
-}
 
 // Change Authentication PIN Form
